requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data processing
numpy>=1.23.0
pandas>=1.5.0
Pillow>=9.0.0

# Metrics
scikit-learn>=1.2.0

# Utils
tqdm>=4.65.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0